# docs/swagger.yaml
basePath: /
definitions:
...
  carsBrandsBattle.Brand:
    properties:
      id:
        type: integer
      imageBrand:
        items:
          type: integer
        type: array
      name:
        type: string
    required:
    - imageBrand
    - name
    type: object
  carsBrandsBattle.Score:
    properties:
      battleId:
        type: integer
      id:
        type: integer
      playerScore:
        type: integer
      userId:
        type: integer
    required:
    - battleId
    - playerScore
    - userId
    type: object
...
  carsBrandsBattle.UserInterfaceData:
    properties:
      battleId:
        type: integer
      player1Username:
        type: string
      player2Username:
        type: string
      playerScore1:
        type: integer
      playerScore2:
        type: integer
      score1Id:
        type: integer
      score2Id:
        type: integer
    type: object
  carsBrandsBattle.UserStatistics:
    properties:
      totalScore:
        type: integer
      userId:
        type: integer
      userName:
        type: string
    type: object
  handler.errorResponse:
    properties:
      message:
        type: string
    type: object
...
  handler.getAllBrandResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/carsBrandsBattle.Brand'
        type: array
    type: object
  handler.getAllScoreResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/carsBrandsBattle.Score'
        type: array
    type: object
  handler.getGeneralUserInterfaceDataResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/carsBrandsBattle.UserInterfaceData'
        type: array
    type: object
  handler.getGeneralUserStatisticsByScoreResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/carsBrandsBattle.UserStatistics'
        type: array
    type: object
...
  handler.statusResponse:
    properties:
      status:
        type: string
    type: object
host: localhost:8000
info:
  contact: {}
  description: API Server for Car Brands Battle Application
  title: Car Brands Battle App API
  version: "1.0"
paths:
...
  /api/brands:
    get:
      consumes:
      - application/json
      description: get all brands
      operationId: get-all-brands
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.getAllBrandResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get All Brands
      tags:
      - brands
    post:
      consumes:
      - application/json
      description: create brand
      operationId: create-brand
      parameters:
      - description: brand info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/carsBrandsBattle.Brand'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create Brand
      tags:
      - brands
  /api/brands/{id}:
    delete:
      consumes:
      - application/json
      description: delete brand
      operationId: delete-brand
      parameters:
      - description: brand id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete Brand
      tags:
      - brands
    get:
      consumes:
      - application/json
      description: get brand by id
      operationId: get-brand-by-id
      parameters:
      - description: brand id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/carsBrandsBattle.Brand'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Brand By Id
      tags:
      - brands
    put:
      consumes:
      - application/json
      description: update brand
      operationId: update-brand
      parameters:
      - description: brand id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update Brand
      tags:
      - brands
  /api/brands/random:
    get:
      consumes:
      - application/json
      description: get random brand
      operationId: get-random-brand
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/carsBrandsBattle.Brand'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Random Brand
      tags:
      - brands
  /api/scores:
    get:
      consumes:
      - application/json
      description: get all scores
      operationId: get-all-scores
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.getAllScoreResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get All Scores
      tags:
      - scores
  /api/scores/{id}:
    delete:
      consumes:
      - application/json
      description: delete score
      operationId: delete-score
      parameters:
      - description: score id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete Score
      tags:
      - scores
    get:
      consumes:
      - application/json
      description: get score by id
      operationId: get-score-by-id
      parameters:
      - description: score id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/carsBrandsBattle.Score'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Score By Id
      tags:
      - scores
    put:
      consumes:
      - application/json
      description: update score
      operationId: update-score
      parameters:
      - description: score id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update Score
      tags:
      - scores
  /api/user-interface-data:
    get:
      consumes:
      - application/json
      description: get user interface data
      operationId: get-user-interface-data
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.getGeneralUserInterfaceDataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get General User Interface Data
      tags:
      - userInterfaceData
  /api/user-statistics/by-score:
    get:
      consumes:
      - application/json
      description: get user statistics by score
      operationId: get-user-statistics
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.getGeneralUserStatisticsByScoreResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get General User Statistics By Score
      tags:
      - userStatistics
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"